refactor(useAvailableCouponList): 쿠폰 code가 아닌 종류에 따라 이용가능성 분류, 쿠폰 util 함수 파일 분리
